[Android] Fix for FRE flow
Chromium change:
https://chromium.googlesource.com/chromium/src/+/eb43b5f66a3acb7e5d05b7dccc8fce6f2f7a5c2d

[AuthTab] Fix the broken FRE flow

AuthTab has a bug of always being canceled when FRE steps in, because
activity result was not being returned due to FRE activity running in a
new task. Subsequent AuthTab activity was also created in a new task
when launched by PendingIntent mechanism. The expected behavior for
AuthTab (CCT in general, for that matter) is for all the activities to
run on the same task as the client app but this is already in a broken
state.

To fix not only AuthTab but also CCT, this CL launches FRE and the next
activity in the same task as that of the client app. This is implemented
by:

- defining a new param |inSameTask| to FirstRunSequencer#launch
- using Intent (as opposed to PendingIntent) object to run the activity
after FRE completion.

Now the change in effect for AuthTab, while CCT is behind a new feature
flag enabled by default so it can switched off upon discovery of
unexpected issues.

Bug: crbug.com/381107767